Flatten exports and rename index.ts files in subdirs
