+++ cloud/wfm/src/executor/InstanceRepository.ts
@@ -1,4 +1,6 @@
-import {ProcessInstance} from '../process-instance';
-export interface InstanceRepository {
+import ProcessInstance from '../process-instance/ProcessInstance';
+interface InstanceRepository {
   save(instance: ProcessInstance): Promise<ProcessInstance>;
 }
+
+export default InstanceRepository;
