+++ cloud/wfm/src/task/BaseTask.ts
@@ -1,6 +1,6 @@
 import { EventEmitter } from 'eventemitter3';
-import {Result} from '../result';
-import { Task, TaskEventData, TaskStatus } from './index';
+import Result from '../result/Result';
+import Task, {TaskEventData, TaskStatus} from './Task';
 
 /**
  * Base implementation for {@link Task}s
