+++ cloud/wfm/test/Process/ProcessImpl.ts
@@ -1,4 +1,4 @@
-import ProcessImpl from '../../src/process';
+import ProcessImpl from '../../src/process/ProcessImpl';
 import suite from './index';
 
 describe('ProcessImpl', function() {
