+++ cloud/wfm/test/ProcessInstance/InstanceImpl.ts
@@ -1,4 +1,4 @@
-import Instance from '../../src/process-instance';
+import Instance from '../../src/process-instance/ProcessInstanceImpl';
 import suite from './index';
 
 describe('InstanceImpl', function() {
