+++ cloud/wfm/test/Result/BooleanResult.ts
