+++ cloud/wfm/test/Result/index.ts
@@ -1,4 +1,4 @@
-import {Result} from '../../src/result';
+import Result from '../../src/result/Result';
 function suite(resultFactory: () => Result) {
   describe('Result', function() {
     it('contains a type value');
